requirements.txt
ansimarkup==1.5.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-default-group==1.2.2
click-repl==0.1.6
colorama==0.4.4
compoundfiles==0.3
compressed-rtf==1.0.6
dicttoxml==1.7.4
gron==1.1.4
halo==0.0.31
idna==2.10
Jinja2==2.11.3
log-symbols==0.0.14
MarkupSafe==1.1.1
more-itertools==8.6.0
msg-parser==1.2.0
olefile==0.46
prompt-toolkit==3.0.14
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
spinners==0.0.24
structlog==20.2.0
-e git+git@github.com:sublime-security/sublime-cli.git@236675f8775caaf5af2f802cbb974f7754df60dd#egg=sublime
termcolor==1.1.0
urllib3==1.26.3
wcwidth==0.2.5
websockets==8.1
